Moving methods from the colorManager mixin to the colorService.
