+++ app/pages/EstudoView.vue
@@ -49,10 +49,10 @@
           v-for="(secao, indexSecao) in estudo.secoes"  
           :key="secao.id"
           row wrap>
-          <v-layout column :id="secao.id" :style="'background-color:' + assessZebraBG(indexSecao) + ';'">
+          <v-layout column :id="secao.id" :style="'background-color:' + $colorsService.assessZebraBG(indexSecao) + ';'">
             <v-flex xs12 v-if="secao.name !== null && secao.name !== undefined">
               <div
-                :class="'display-2-obs text-xs-center pt-5 pb-3 font-weight-bold ' + assessZebraTitle(indexSecao)">
+                :class="'display-2-obs text-xs-center pt-5 pb-3 font-weight-bold ' + $colorsService.assessZebraTitle(indexSecao)">
                 {{ secao.name }}
               </div>
             </v-flex>
@@ -77,7 +77,7 @@
                   </v-layout>
                   <v-layout v-else-if="card.type && card.type == 'headline'"
                     justify-center pt-5 pb-3
-                    :class="'display-2-obs font-weight-bold ' + assessZebraTitle(indexSecao)"
+                    :class="'display-2-obs font-weight-bold ' + $colorsService.assessZebraTitle(indexSecao)"
                     v-html="card.title.fixed">
                   </v-layout>
                   <flpo-story-card-autofill
@@ -116,7 +116,7 @@
     </v-container>
     <!-- Navegação lateral em dots pelos dimensoes e seções -->
     <flpo-dot-nav :sections="estudo.secoes"></flpo-dot-nav>
-    <!-- <v-container fluid ma-0 pa-5 :style="'background-color:' + assessZebraBG(1) + ';'">
+    <!-- <v-container fluid ma-0 pa-5 :style="'background-color:' + $colorsService.assessZebraBG(1) + ';'">
       <v-layout row wrap text-xs-center pb-5>
         <div class="flex display-1-obs">Realização</div>
       </v-layout>
