+++ app/pages/LocalidadeView.vue
@@ -137,10 +137,10 @@
           v-for="(secao, indexSecao) in sections"  
           :key="secao.id"
           row wrap>
-          <v-layout column :id="secao.id" :style="'background-color:' + assessZebraBG(indexSecao) + ';'">
+          <v-layout column :id="secao.id" :style="'background-color:' + $colorsService.assessZebraBG(indexSecao) + ';'">
             <v-flex xs12>
               <div
-                :class="'display-2-obs pt-5 pb-3  ml-5 pl-3 font-weight-bold ' + assessZebraTitle(indexSecao)">
+                :class="'display-2-obs pt-5 pb-3  ml-5 pl-3 font-weight-bold ' + $colorsService.assessZebraTitle(indexSecao)">
                 {{ secao.name }}
               </div>
             </v-flex>
@@ -174,7 +174,7 @@
                     </v-layout>
                     <v-layout v-else-if="card.type && card.type == 'headline'"
                       pt-5 pb-3 ml-5 pl-2
-                      :class="'display-2-obs font-weight-bold ' + assessZebraTitle(indexSecao)"
+                      :class="'display-2-obs font-weight-bold ' + $colorsService.assessZebraTitle(indexSecao)"
                       v-html="card.title.fixed">
                     </v-layout>
                     <flpo-story-card-autofill
