+++ app/test/mixins/service/colorManager.test.js
@@ -112,4 +112,74 @@ describe('ColorManager', () => {
     let result = wrapper.vm.$colorsService.getColorScale('RdYlGn', 'divergent', 'asc', 10);
     expect(result).toEqual(["#a50026", "#d73027", "#f46d43", "#fdae61", "#fee08b", "#d9ef8b", "#a6d96a", "#66bd63", "#1a9850", "#006837"]);
   })
+
+  test('Verifica se pega corretamente o valor da classe primária de BG zebrado', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    wrapper.vm.changeTheme('default'); // Sets initial theme
+    let result = wrapper.vm.$colorsService.assessZebraBG(0);
+    expect(result).toEqual("#EFEFEF");
+  })
+
+  test('Verifica se pega corretamente o valor da classe secundária de BG zebrado', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.assessZebraBG(1);
+    expect(result).toEqual("#e0e0e0");
+  })
+
+  test('Quando o tema for escuro, deve retornar a classe correspondente', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.getClassIfIsDark(null, 0);
+    expect(result).toEqual("theme--dark");
+  })
+
+  test('Quando não houver o dado da cor do bg e nenhum índice for informado, volta classe em branco.', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.getClassIfIsDark(null, null);
+    expect(result).toEqual("");
+  })
+
+  test('Quando o tema for escuro, o título do zebrado deve vir branco', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    wrapper.vm.$vuetify.theme.background = '#000000';
+    let result = wrapper.vm.$colorsService.assessZebraTitle(0);
+    expect(result).toEqual("white--text");
+  })
+
+  test('Quando o tema for escuro, a propriedade de cor do zebrado deve vir branca', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    wrapper.vm.$vuetify.theme.background1 = colors.grey.darken4;
+    wrapper.vm.$vuetify.theme.background2 = colors.grey.lighten2;
+    let result = wrapper.vm.$colorsService.assessZebraTitleColor(0);
+    expect(result).toEqual("white");
+  })
+
+  test('Quando o tema for claro, a propriedade de cor do zebrado deve vir preta', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.assessZebraTitleColor(1);
+    expect(result).toEqual("black");
+  })
+
+  test('Quando o tema for escuro, a propriedade de cor translúcida do zebrado deve vir branca com opacidade 0.7', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.assessZebraTitleColor(0, 0.7);
+    expect(result).toEqual("rgba(255, 255, 255, 0.7)");
+  })
+
+  test('Quando o tema for claro, a propriedade de cor translúcida do zebrado deve vir preta com opacidade 0.7', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.assessZebraTitleColor(1, 0.7);
+    expect(result).toEqual("rgba(0, 0, 0, 0.7)");
+  })
+
+  test('Quando o tema for escuro, a propriedade de cor do eixo deve vir branca', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.assessZebraAxesColor(0);
+    expect(result).toEqual("white");
+  })
+
+  test('Quando o tema for claro, a propriedade de cor do eixo deve vir cinza', () => {
+    const wrapper = mount(FLPOSobreLayout, { sync: false })
+    let result = wrapper.vm.$colorsService.assessZebraAxesColor(1);
+    expect(result).toEqual(colors.grey.base);
+  })
 })
